+++ packages/omero-screen-napari/src/omero_screen_napari/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.4"
+__version__ = "0.2.5"
 
 from omero_screen.config import set_env_vars